Develop

+++ core/database.php
@@ -10,6 +10,7 @@ use \core\functions;
 /**
  * @class Database
  */
+
 class database
 {
 
@@ -100,13 +101,13 @@ class database
     {
         return $this->_connection->prepare($sql);
     }
-     //Procesar consulta, sql=consulta,return=devolver resultados o solo true o false si se ejecuto la operacion
-    private function consulta($sql, $return)
+    //Procesar consulta, sql=consulta,return=devolver resultados o solo true o false si se ejecuto la operacion
+    private function consulta($sql, $return, $delete_cache = true)
     {
         /*foreach ($this->llamadas as $key => $ll) {
-            if($ll['consulta']==$sql){
-                return $ll['resultado'];
-            }
+        if($ll['consulta']==$sql){
+        return $ll['resultado'];
+        }
         }*/
         //$ll=array('consulta'=>$sql);
         //$t=microtime(true)*1000;
@@ -115,8 +116,10 @@ class database
             $query->execute();
             if ($return) {
                 $rows = $query->fetchAll();
-            }else{
-                cache::delete_cache();
+            } else {
+                if ($delete_cache) {
+                    cache::delete_cache();
+                }
             }
 
         } catch (\PDOException $e) {
@@ -134,7 +137,7 @@ class database
             }
         }
         //$ll['resultado']=$rows;
-       // $ll['tiempo']=(microtime(true)*1000)-$t;
+        // $ll['tiempo']=(microtime(true)*1000)-$t;
         //$this->llamadas[]=$ll;
 
         return $rows;
@@ -142,15 +145,15 @@ class database
 
     public function get_last_insert_id()
     { //ultimo elemento insertado
-        return (int)$this->_connection->lastInsertId();
+        return (int) $this->_connection->lastInsertId();
     }
 
     public function get($table, $idname, $where, $condiciones = array(), $select = "")
     {
         if ($select == "") {
             $select = "*";
-        }elseif($select=='total'){
-            $select=$idname;
+        } elseif ($select == 'total') {
+            $select = $idname;
         }
 
         $sql = "SELECT " . $select . " FROM " . self::$_prefix . $table;
@@ -189,7 +192,7 @@ class database
         return $row;
     }
 
-    public function insert($table, $idname, $insert)
+    public function insert($table, $idname, $insert, $delete_cache = true)
     { //consulta insert
         $valor_primario = "";
         $image          = array();
@@ -213,9 +216,9 @@ class database
             $sql .= ($value == "true" || $value == "false") ? $value : "'" . str_replace("'", "\\'", $value) . "'";
         }
         $sql .= ")";
-        $row = $this->consulta($sql, false);
-        if($row){
-            $last_id=$this->get_last_insert_id();
+        $row = $this->consulta($sql, false, $delete_cache);
+        if ($row) {
+            $last_id = $this->get_last_insert_id();
             if (count($image) > 0) {
                 $this->process_image($image, $table, $idname, $last_id);
             }
@@ -227,7 +230,7 @@ class database
         return $row;
     }
 
-    public function update($table, $idname, $set, $where)
+    public function update($table, $idname, $set, $where, $delete_cache = true)
     { //consulta update
         $set   = self::process_multiple($set);
         $image = array();
@@ -259,8 +262,8 @@ class database
         }
         $sql .= ") ";
         if (count($where) > 0) {
-            $row = $this->consulta($sql, false);
-            if($row){
+            $row = $this->consulta($sql, false, $delete_cache);
+            if ($row) {
                 if (count($image) > 0) {
                     $this->process_image($image, $table, $idname, $where[$idname]);
                 }
@@ -275,7 +278,7 @@ class database
         }
     }
 
-    public function delete($table, $idname, $where)
+    public function delete($table, $idname, $where, $delete_cache = true)
     { //consulta delete
         $sql = "DELETE FROM " . self::$_prefix . $table;
         $sql .= " WHERE (TRUE";
@@ -284,7 +287,7 @@ class database
         }
         $sql .= ")";
         if (count($where) > 0) {
-            $row = $this->consulta($sql, false);
+            $row = $this->consulta($sql, false, $delete_cache);
             image::delete($table, '', $where[$idname]);
             file::delete($table, '', $where[$idname]);
             return $row;
